# Backend/Backend/Controllers/CommunityController.cs
using Backend.Entities;
using Backend.Services;
using Humanizer;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using System;
using System.IdentityModel.Tokens.Jwt;
using System.Security.Claims;

namespace Backend.Controllers
{
    [ApiController]
    [Route("[controller]")]
    public class CommunityController : ControllerBase
    {
        private readonly AppDbContext _context;
        private readonly ValidationService _validationService;

        public CommunityController(AppDbContext context, ValidationService validationService)
        {
            _context = context;
            _validationService = validationService;
        }

        [HttpGet]
        public async Task<ActionResult<IEnumerable<CommunityResponseDTO>>> GetCommunities()
        {
            var communityList = await _context.Community.Include(c => c.User).ToListAsync();
            if (communityList == null || communityList.Count == 0)
            {
                return NotFound("No communities found.");
            }

            var userId = HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
            bool isAdmin = HttpContext.User.IsInRole("Admin");

            if (string.IsNullOrEmpty(userId))
            {
                userId = "-1";
            }

            var responseDTO = communityList.Select(c => c.toDto(long.Parse(userId), isAdmin)).ToList();

            return Ok(responseDTO);
        }

        [HttpGet("{id}")]
        public async Task<ActionResult<CommunityResponseDTO>> GetCommunity(long id)
        {
            var community = await _context.Community.FindAsync(id);
            
            if (community == null)
            {
                return NotFound($"Community with ID {id} not found.");
            }

            await _context.Entry(community)
                .Reference(c => c.User)
                .LoadAsync();

            var userId = HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
            bool isAdmin = HttpContext.User.IsInRole("Admin");

            var responseDTO = community.toDto(long.Parse(userId), isAdmin);

            return Ok(responseDTO);
        }

        [Authorize(Roles = "Member,Admin")]
        [HttpPost]
        public async Task<ActionResult<CommunityResponseDTO>> CreateCommunity([FromBody] CommunityDTO communityDto)
        {
            if (!_validationService.ValidateCommunityDTO(communityDto))
            {
                if (communityDto == null)
                {
                    return BadRequest("Invalid community data.");
                }
                return UnprocessableEntity("Invalid community data.");
            }

            var user = await _context.Users.FindAsync(long.Parse(HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub)));
            if (user == null)
            {
                return NotFound("User not found.");
            }

            Community community = new Community
            {
                Name = communityDto.name,
                Description = communityDto.description,
                UserId = user.Id,
                User = user
            };

            community.CreationDate = DateTime.UtcNow;

            _context.Community.Add(community);
            await _context.SaveChangesAsync();

            var userId = HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
            bool isAdmin = HttpContext.User.IsInRole("Admin");

            var responseDTO = community.toDto(long.Parse(userId), isAdmin);

            return CreatedAtAction(nameof(GetCommunity), new { id = community }, responseDTO);
        }

        [Authorize(Roles = "Member,Admin")]
        [HttpPut("{id}")]
        public async Task<ActionResult<CommunityResponseDTO>> UpdateCommunity([FromBody] CommunityDTO communityDto, long id)
        {
            if (!_validationService.ValidateCommunityDTO(communityDto))
            {
                if (communityDto == null)
                {
                    return BadRequest("Invalid community data.");
                }
                return UnprocessableEntity("Invalid community data.");
            }

            if (id <= 0)
            {
                return BadRequest("Invalid community ID.");
            }

            Community community = await _context.Community.FindAsync(id);
            if (community == null)
            {
                return NotFound($"Community with ID {id} not found.");
            }

            if (!HttpContext.User.IsInRole("Admin") && long.Parse(HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub)) != community.UserId)
            {
               return Forbid();
            }

            community.Name = communityDto.name;
            community.Description = communityDto.description;

            _context.Community.Update(community);
            await _context.SaveChangesAsync();

            await _context.Entry(community)
                .Reference(c => c.User)
                .LoadAsync();

            var userId = HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
            bool isAdmin = HttpContext.User.IsInRole("Admin");

            var responseDTO = community.toDto(long.Parse(userId), isAdmin);

            return Ok(responseDTO);
        }

        [Authorize(Roles = "Member,Admin")]
        [HttpDelete("{id}")]
        public async Task<IActionResult> Delete(long id)
        {
            var entity = await _context.Community.FindAsync(id);
            if (entity == null) return NotFound();

            if (!HttpContext.User.IsInRole("Admin") && long.Parse(HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub)) != entity.UserId)
            {
               return Forbid();
            }

            _context.Community.Remove(entity);
            await _context.SaveChangesAsync();
            return NoContent();
        }

        [Authorize(Roles = "Admin")]
        [HttpGet("{community_id}/posts/comments")]
        public async Task<ActionResult<IEnumerable<CommentResponseDTO>>> GetCommentsByCommunityId(long community_id)
        {
            var posts = await _context.Post.Where(p => p.CommunityId == community_id)
                .Select(p => p.Id)
                .ToListAsync();
            if (posts == null || posts.Count == 0)
            {
                return NotFound($"No posts found for community ID {community_id}.");
            }

            var comments = await _context.Comment
                .Where(c => posts.Contains(c.PostId))
                .Include(c => c.User)
                .ToListAsync();
            if (comments == null || comments.Count == 0)
            {
                return NotFound($"No comments found for community ID {community_id}.");
            }

            var userId = HttpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub);
            bool isAdmin = HttpContext.User.IsInRole("Admin");

            var responseDTO = comments.Select(c => c.toDto(long.Parse(userId), isAdmin)).ToList();
            return Ok(responseDTO);
        }
    }
}